--- app/javascript/components/CardOwnership/format-card.js
@@ -1,9 +1,0 @@
-export default (card, ownerships) => ({
-  id: card.id,
-  attributes: {
-    name: card.name,
-    users: {
-      data: ownerships.map((o) => ({ attributes: o.collection.user })),
-    },
-  },
-});
