+++ app/javascript/packs/application.scss
@@ -4,6 +4,7 @@
 @import "../stylesheets/admin/users";
 @import "../stylesheets/sitewide/button";
 @import "../stylesheets/sitewide/toggle-switch";
+@import "../stylesheets/sitewide/icon";
 @import "../stylesheets/cards/card-profile";
 @import "../stylesheets/card-grid/card-grid";
 @import "../stylesheets/dashboard/dashboard";
