+++ app/javascript/stylesheets/sitewide/button.scss
@@ -100,17 +100,6 @@ a.button,
         }
     }
 
-    &--togglable {
-        --button-background-color-lightness:  calc(
-            var(--color-fill-theme-lightness) + (15% * var(--color-lightness-shift-basis))
-        );
-        border: 2px solid var(--button-background-color);
-
-        &.on {
-            --button-background-color-lightness: var(--color-fill-theme-lightness);
-        }
-    }
-
     .hollow-text {
         color: var(--button-background-color);
         transition-duration: 0.3s;
