+++ app/serializers/user_serializer.rb
@@ -1,5 +1,4 @@
 class UserSerializer
   include FastJsonapi::ObjectSerializer
   attributes :name, :id
-
 end
